# backend/backend-main/Swagger_API.yaml
swagger: "2.0"
info:
  description: API from Web Technology project 'Nolo Nolo+'
  version: 1.0.0
  title: NoloNolo+APIs
  termsOfService: http://swagger.io/terms/
  contact:
    email: josephcarrino2203@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: api.server.test
schemes:
- https
- http
paths:
...
  /users/{id}/rentals:
    get:
      tags:
      - Users
      summary: Get the rentals of the user with the specified id
      operationId: getAllRentalOfUserByID
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: Successful Operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Rental'
        "400":
          description: Invalid ID
        "404":
          description: ID not found
...
  /articles/{id}/rentals:
    get:
      tags:
      - Articles
      summary: Get the rentals of the article with the specified id
      operationId: getAllRentalOfArticleByID
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: Successful Operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Rental'
        "400":
          description: Invalid ID
        "404":
          description: ID not found
...
  /staff/{id}/rentals:
    get:
      tags:
      - Staff
      summary: Get the rentals of the employee with the specified id
      operationId: getAllRentalOfEmployeeByID
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: Successful Operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Rental'
        "400":
          description: Invalid ID
        "404":
          description: ID not found
...
  /rentals:
    get:
      tags:
      - Rentals
      summary: Get a list of all rentals
      operationId: getAllRentals
      parameters: []
      responses:
        "200":
          description: Succesful Operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Rental'
        "400":
          description: Server Error
    post:
      tags:
      - Rentals
      summary: Add a new rental (ID given server-side)
      operationId: createRental
      parameters:
      - in: body
        name: RentalObj
        required: true
        schema: {}
      responses:
        "200":
          description: Operation Successful
  /rentals/{id}:
    get:
      tags:
      - Rentals
      summary: Get a rental (ID given server-)
      operationId: getRentalByID
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: Operation Successful
          schema: {}
        "400":
          description: Invalid ID
        "404":
          description: ID not found
    post:
      tags:
      - Rentals
      summary: Modify a rental
      operationId: modRentalByID
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - in: body
        name: rentalObj
        required: true
        schema: {}
      responses:
        "200":
          description: Operation Successful
        "400":
          description: Invalid ID
        "404":
          description: ID not found
    delete:
      tags:
      - Rentals
      summary: Delete the rental with the specified id
      operationId: deleteRentalByID
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: Successful Operation
        "400":
          description: Invalid ID
        "404":
          description: ID not found
...
  /categories/{id}/rentals:
    get:
      tags:
      - Categories
      summary: Get the rentals of the category with the specified id
      operationId: getCatgoryRentals
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: Successful Operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Rental'
        "400":
          description: Invalid ID
        "404":
          description: ID not found
...
definitions:
...
  Rental:
    type: object
    required:
    - date_end
    - date_start
    - estimated
    - functionary
    - id
    - object_id
    - state
    - user
    properties:
      id:
        type: string
        example: 614dc8a650f8cab4eb8c8c12
      user:
        $ref: '#/definitions/User/properties/username'
      functionary:
        $ref: '#/definitions/Employee/properties/username'
      object_id:
        $ref: '#/definitions/Article/properties/id'
      date_start:
        type: string
        format: yyyy-mm-dd
        example: 2021-11-14
      date_end:
        type: string
        format: yyyy-mm-dd
        example: 2021-11-18
      estimated:
        type: number
        example: 20.67
      state:
        type: string
        example: Pending
...